[T] constants: move solutions data
