+++ src/constants/SOLUTIONS.js
